3-text-extractor/backend/requirements.txt
fastapi
openai
psycopg2-binary
pydantic
pydantic-settings
uvicorn
